==> analyses/0.3_sample-articles-for-scoping.Rmd <==
---
title: "0.3_sample-articles-for-scoping"
author: "Devi Veytia"
date: "2024-04-25"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Set up -- run before anything else

```{r set up libraries}
## Load libraries
library(dplyr)
library(dbplyr)
library(R.utils)
library(RSQLite)
library(ggplot2)
library(tidyverse)

# hack for error with coord_cartesian in ggplot
# https://stackoverflow.com/questions/78227158/big-issue-with-ggplot2-r-and-coord-cartesian
if(!exists("obj_is_vector")){
  obj_is_vector = function(x){TRUE}
}

```

```{r set the seed}
addTaskCallback(function(...) {set.seed(123);TRUE})
```


# Sample the articles - for coding scoping 

sampling approach for selecting articles:
- sample evenly from three different mitigation ORO branches: MRE, increase efficiency, CDR
- sample only predicted relevance > 0.8 because the purpose of this exercise is to practise the codebook, want to try to target articles most likely to be relevant.


```{r sample articles for scoping coding}
## Set up inputs to determine how many articles to sample
setSize <- 200 # write the files in batches to not overwhelm sysrev
nRecsTotal = 2000 # No of articles to sample in total



## connect to database
p1_db <- RSQLite::dbConnect(RSQLite::SQLite(),
                      here::here("data","sqlite-databases","product1.sqlite"),
                                 create=FALSE)


# Collect table of unique references and categories (impact nature + different ORO types) to sample from
relevantdedups <- tbl(p1_db, "uniquerefs") %>%
  select(analysis_id, type, title, journal, year, volume, number, abstract,doi, # author, 
         issn, keywords) %>%
  collect()

pred_OROmitigation <- tbl(p1_db, "pred_oro_any_mitigation") %>%  # mitigation ORO
  # # Filter to relevance >=0.8
  # filter(0.8 <= `oro_any.M_Renewables - mean_prediction`|
  #        0.8 <= `oro_any.M_Increase_efficiency - mean_prediction`|
  #        0.8 <= `oro_any.M_CO2_removal_or_storage - mean_prediction`) %>%
  select(analysis_id, `oro_any.M_Renewables - mean_prediction`, `oro_any.M_Increase_efficiency - mean_prediction`,
         `oro_any.M_CO2_removal_or_storage - mean_prediction`) %>%
  collect()

# For simplicity, remove articles where multiple OROs are predicted relevant
pred_OROmitigation <- pred_OROmitigation %>%
  mutate(n_OROs = rowSums(.[2:4] > 0.8)) %>%
  filter(n_OROs == 1) %>%
  select(-c(n_OROs))


# sample data frame evenly for each ORO_type
sampleDf <- reshape2::melt(pred_OROmitigation, id.vars = "analysis_id", variable.name = "ORO_type",
                           value.name = "mean_prediction")
sampleDf <- sampleDf %>%
  filter(0.8 <= mean_prediction) %>%
  group_by(ORO_type) %>%
  slice_sample(n = round(nRecsTotal/3)) %>% 
  ungroup() %>%
  inner_join(relevantdedups, by="analysis_id") # Join with metadata

# shuffle the row so they're in random order
sampleDf <- sampleDf[sample(1:nrow(sampleDf), nrow(sampleDf), replace = FALSE),]
sampleDf <- sampleDf[1:nRecsTotal,] # if there is an extra due to rounding remove

# Write out a file that keeps a record of which articles were sampled to make a lookup table
dbWriteTable(p1_db, "scopingCodingSample", sampleDf, overwrite = FALSE, append = FALSE)
dbDisconnect(p1_db) # Disconnect the database


## Write articles in batches to .ris files
startInd <- seq(1, nrow(sampleDf), by = setSize) # the start id of the set
endInd <- startInd+setSize-1
endInd <- ifelse(endInd > nrow(sampleDf), nrow(sampleDf), endInd)

for(i in 1:length(startInd)){
  temp <- sampleDf[startInd[i]:endInd[i],]
  temp <- as.data.frame(temp)
  revtools::write_bibliography(temp, filename = here::here(
        "data/derived-data/coding-scoping/",paste0("coding_scoping_set_",i,".ris")))
}

```



# Analyse scoping results

```{r set common factoring levels}
scopeOROFactorLevels <- c("MRE - marine energy","MRE - located",
                          "Efficiency - ships","Efficiency - fish",
                          "CDR - BC", "CDR - CCS", "CDR - Macroalgae", "CDR - OAE","CDR - OIF")
```

```{r load in scoping results}

# # Load in metadata
p1_db <- RSQLite::dbConnect(RSQLite::SQLite(),
                      here::here("data","sqlite-databases","product1.sqlite"),
                                 create=FALSE)
sampleDf <- tbl(p1_db, "scopingCodingSample") %>%
  collect()

dbDisconnect(p1_db)

# Load in scoping results
scope_df <- readxl::read_excel(
  here::here("data/derived-data/coding-scoping/product-1-coding-scoping_2024-06-21.xlsx"),
  sheet = "Article Answers"
)


# Join scoping results to metadata
analysis_id_lookup <- fuzzyjoin::stringdist_inner_join(scope_df %>% select(`Article ID`, Title) %>%
                                                         rename(title = Title), 
                                      sampleDf %>% 
                                        select(analysis_id, title, abstract, doi, year),
                                      by ="title", max_dist=2, ignore_case = TRUE)

if(nrow(analysis_id_lookup) > nrow(scope_df)){ # if there are any dups from join, extract the most complete reference
  analysis_id_lookup <- revtools::extract_unique_references(analysis_id_lookup, analysis_id_lookup$`Article ID`)
}

analysis_id_lookup <- analysis_id_lookup %>% # Keep title from original metadata and format/clean data frame
  rename(title = title.y) %>%
  select(`Article ID`, analysis_id)

```

```{r barplot summarising counts for all variables}

## Loop through all the variables and count the number of mentions of each label
variables <- colnames(scope_df)[6:13]

for(v in 1:length(variables)){
  
  # variable column of interest
  var <- unlist(scope_df[,variables[v]])
  
  # Get a vector of all the unique labels
  labels <- unique(var)
  labels <- labels[!is.na(labels)]
  labels <- unlist(strsplit(labels, "[|||]"))
  labels <- labels[labels != ""]
  labels <- unique(labels)
  labels <- sort(labels)
  
  # For each label, count the occurrences
  lab_count <- rep(NA, length(labels))
  for(l in 1:length(labels)){
    lab_count[l] <- sum(grepl(labels[l], var))
  }
  
  # save results
  tmp_df <- data.frame(
      variable = rep(variables[v], length(labels)),
      label = labels,
      label_count = lab_count
    )
  
  if(v==1){
    label_counts <- tmp_df
  }else{
    label_counts <- rbind(label_counts, tmp_df)
      
  }
  
  # Clean up environment from temporary variables
  if(v==length(variables)){
    rm(tmp_df, lab_count, labels, var)
  }
  
}

label_counts$proportion <- label_counts$label_count/nrow(scope_df)



## Calculate the projected number of relevant articles based on scoping

# Get the projected number of relevant articles for the different types of mitigation OROs
p1_db <- RSQLite::dbConnect(RSQLite::SQLite(),
                      here::here("data","sqlite-databases","product1.sqlite"),
                                 create=FALSE)

pred_OROmitigation_sums <- tbl(p1_db, "pred_oro_any_mitigation") %>%  
  filter(0.5 <= `oro_any.M_Renewables - mean_prediction`|
         0.5 <= `oro_any.M_Increase_efficiency - mean_prediction`|
         0.5 <= `oro_any.M_CO2_removal_or_storage - mean_prediction`) %>%
  select(analysis_id, `oro_any.M_Renewables - mean_prediction`, `oro_any.M_Increase_efficiency - mean_prediction`,
         `oro_any.M_CO2_removal_or_storage - mean_prediction`) %>%
  summarise(
    all_mit = n(), 
    MRE = sum(0.5 <= `oro_any.M_Renewables - mean_prediction`),
    Efficiency = sum(0.5 <= `oro_any.M_Increase_efficiency - mean_prediction`),
    CDR = sum(0.5 <= `oro_any.M_CO2_removal_or_storage - mean_prediction`)
  ) %>%
  collect()

dbDisconnect(p1_db)

# Use the relevant sums to weight the counts
# If it deals with a specific ORO type, weight by that ORO prediction, otherwise just give as a proportion of all mitigation values
label_counts$projected_n <- rep(NA, nrow(label_counts))

label_counts$projected_n[
  grep("MRE", label_counts$label, ignore.case = FALSE)] <- label_counts$proportion[
    grep("MRE", label_counts$label, ignore.case = FALSE)]*pred_OROmitigation_sums$MRE

label_counts$projected_n[
  grep("Efficiency", label_counts$label, ignore.case = FALSE)] <- label_counts$proportion[
    grep("Efficiency", label_counts$label, ignore.case = FALSE)]*pred_OROmitigation_sums$Efficiency

label_counts$projected_n[
  grep("CDR", label_counts$label, ignore.case = FALSE)] <- label_counts$proportion[
    grep("CDR", label_counts$label, ignore.case = FALSE)]*pred_OROmitigation_sums$CDR

label_counts$projected_n[is.na(label_counts$projected_n)] <- label_counts$proportion[is.na(label_counts$projected_n)]*pred_OROmitigation_sums$all_mit

# round
label_counts$projected_n <- round(label_counts$projected_n)


# Format for plotting

# factoring variables
label_counts$variable_f <- factor(label_counts$variable, levels = variables)

# Add a space between variable groups
empty_bar <- 1 # Number of empty bars to space out groups
to_add <- data.frame( matrix(NA, empty_bar*nlevels(label_counts$variable_f), ncol(label_counts)) )
colnames(to_add) <- colnames(label_counts)
to_add$variable_f <- rep(levels(label_counts$variable_f), each=empty_bar)
label_counts2 <- rbind(label_counts, to_add)
label_counts2 <- label_counts2 %>% arrange(variable_f)
label_counts2$id <- seq(1, nrow(label_counts2))


# # Get the name and the y position of each label
# label_data <- label_counts2
# number_of_bar <- nrow(label_data)
# angle <- 90 - 360 * (label_data$id-0.5) /number_of_bar # I substract 0.5 because the letter must have the angle of the center of the bars. Not extreme right(1) or extreme left (0)
# label_data$hjust <- ifelse( angle < -90, 1, 0)
# label_data$angle <- ifelse(angle < -90, angle+180, angle)

# prepare a data frame for base lines
base_data <- label_counts2 %>% 
  group_by(variable_f) %>% 
  summarize(start=min(id), end=max(id) - empty_bar) %>% 
  rowwise() %>% 
  mutate(title=mean(c(start, end)))

mround <- function(x,base){
	base*round(x/base)
}

y_axis_labs <- data.frame(
  x=rep(-0.3, 5), 
  y = round(seq(0, max(label_counts2$proportion, na.rm=T), length.out = 5), digits = 2),
  label = round(seq(0, max(label_counts2$proportion, na.rm=T), length.out = 5)*nrow(scope_df), digits = 0)
)
y_axis_labs$label <- as.character(mround(y_axis_labs$label, 5))

## Plot
label_counts_ggp <- ggplot(label_counts2, aes(x=as.factor(id), y=proportion, fill=variable_f)) + 
  # Plot bars
  geom_bar(stat="identity", alpha=0.5) +
  # Add labels for projected numbers of articles
  geom_text(aes(label=formatC(projected_n, format = "d", big.mark = ",")), color="black", fontface="bold",alpha=0.6, size=2.5, 
            hjust=0, angle= 90, vjust = 0) +
  
  # Add x labels
  geom_text(aes(label = str_wrap(label, width = 35)), y = -0.01, hjust = 1, vjust=0.7, angle = 90, colour = "black", alpha=0.8, size=2.5)+
  
  # add group lines
  geom_segment(data=base_data, aes(x = start, y = -0.35, xend = end, yend = -0.35), colour = "black", alpha=0.8, linewidth=0.6 , inherit.aes = FALSE )  +
  geom_text(data=base_data, aes(x = title, y = -0.36, label=variable_f), hjust=1, angle = 30, colour = "black", alpha=0.8, size=2.5, fontface="bold", inherit.aes = FALSE)+
  
  # add y axis
  geom_segment(x = -0.2, y = 0, xend = -0.2, yend = max(label_counts2$proportion, na.rm=T), colour = "black", alpha=0.8, linewidth=0.3 , inherit.aes = FALSE )+
  geom_text(data = y_axis_labs, aes(x=x, y = y, label=label),inherit.aes = FALSE, size = 2, hjust = 1)+
  geom_text(x=-2, y= max(label_counts2$proportion, na.rm=T)/2, label = "N scoped articles", angle = 90,
            vjust = 0, colour = "black", alpha=0.8, size=2.5, inherit.aes = FALSE)+
 
  # Limits and other plot aesthetics
  ylim(c(-0.5,max(label_counts2$proportion, na.rm=T))) +
  coord_cartesian(clip = "off")+
  theme_minimal() +
  theme(
    legend.position = "none",
    axis.text = element_blank(),
    axis.title = element_blank(),
    panel.grid = element_blank(),
    plot.margin = unit(c(0.2, 0, 1.5, 1), "cm") 
  ) 

label_counts_ggp
 
ggsave(
  filename = here::here("figures/2024-07-01_expert-panel-meeting/scoping-counts-barplot.pdf"),
  plot = label_counts_ggp,
  height = 6.5, width = 6, units="in"
)
```



```{r validation -- compare predicted oro type vs hand coded}

## Load in predictions from systematic map and the metadata from the scoped articles
p1_db <- RSQLite::dbConnect(RSQLite::SQLite(),
                      here::here("data","sqlite-databases","product1.sqlite"),
                                 create=FALSE)
sampleDf <- tbl(p1_db, "scopingCodingSample") %>% collect()
dbDisconnect(p1_db)


## Format scoping results for the ORO type variable
# Get a vector of all the unique labels
labels <- unique(unlist(scope_df[,"ORO type"]))
labels <- labels[!is.na(labels)]
labels <- unlist(strsplit(labels, "[|||]"))
labels <- labels[labels != ""]
labels <- unique(labels)
labels <- sort(labels)
labels

# For each article, identify if it has been coded yes for a given label
label_mat <- matrix(nrow = nrow(scope_df), ncol = length(labels), dimnames = list(c(scope_df$`Article ID`), labels))
for(i in 1:nrow(scope_df)){
  for(l in 1:length(labels)){
    label_mat[i,l] <- sum(grepl(labels[l], scope_df$`ORO type`[i]))
  }
}
label_mat <- as.data.frame(label_mat)
label_mat$`Article ID` <- scope_df$`Article ID`


# Join with metadata to get the title, which can be matched to get the analysis id
label_mat <- label_mat %>%
  inner_join(scope_df %>% select(`Article ID`, Title))
label_mat <- label_mat %>% rename(title = Title)

label_df <- fuzzyjoin::stringdist_inner_join(label_mat, 
                                      sampleDf %>% 
                                        select(analysis_id, title, abstract, doi, year),
                                      by ="title", max_dist=2, ignore_case = TRUE)

if(nrow(label_df) > nrow(label_mat)){ # if there are any dups from join, extract the most complete reference
  label_df <- revtools::extract_unique_references(label_df, label_df$`Article ID`)
}

label_df <- label_df %>% # Keep title from original metadata and format/clean data frame
  select(-c(`Article ID`, title.x, n_duplicates)) %>%
  rename(title = title.y)

label_df_long <- reshape2::melt(label_df, id.vars = c("analysis_id","title","abstract","doi"), 
                                variable.name = "ORO type", value.name = "coded_value")

label_df_long <- label_df_long %>%
  filter(coded_value == 1) %>%
  select(-c(coded_value)) %>%
  rename(Coded_ORO_type = `ORO type`)


## Join with mean prediction for the different ORO types
ORO_type_validation <- label_df_long %>%
  inner_join(sampleDf %>% select(analysis_id, ORO_type, mean_prediction))
ORO_type_validation$ORO_type <- factor(ORO_type_validation$ORO_type,
                                       levels = c("oro_any.M_Renewables - mean_prediction", 
                                                  "oro_any.M_Increase_efficiency - mean_prediction",
                                                  "oro_any.M_CO2_removal_or_storage - mean_prediction"),
                                       labels = c("MRE","Increase efficiency","CDR"))
ORO_type_validation$Coded_ORO_type <- factor(ORO_type_validation$Coded_ORO_type,
                                             levels = c("MRE - marine energy","MRE - located",
                                                        "Efficiency - ships","Efficiency - fish",
                                                        "CDR - BC", "CDR - CCS", "CDR - Macroalgae", "CDR - OAE","CDR - OIF"))



## Plot sankey diagram
library(ggsankey)

temp <- ORO_type_validation %>%
  make_long(ORO_type, Coded_ORO_type) 
  

tempTab <- temp %>%
  group_by(node) %>%
  tally()

temp <- merge(temp, tempTab, by = "node", all.x=TRUE)

## PLOT THE SANKEY DIAGRAM
oroTypeValidationSankey <- temp %>%
  ggplot(aes(x=x, next_x=next_x, node=node, next_node=next_node, 
             fill=factor(node), label = paste0(node, "\n n=", n))) +
  geom_sankey(flow.alpha=0.5, node.color = "black", show.legend = FALSE)+
  geom_sankey_label(size = 2.5, color = "black", fill= "white", hjust = 0)+
  scale_fill_viridis_d()+
  # scale_fill_manual(
  #   values = c(
  #     'Blue carbon' = "#1f78b4",
  #     'Seagrass' = '#b2df8a',
  #     'Salt marsh' = '#fb9a99',
  #     "Other" = "darkgrey",
  #     "Mangrove" = "#33a02c",
  #     "Macroalgae" = "#cab2d6",
  #     "Coral reef" = "#ff7f00",
  #     "Mitigation" = "#e31a1c"
  #   )
  # )+
  labs(fill = 'Nodes')+
  theme_bw()+
  theme(
    legend.position = "none",
    panel.border = element_blank(),
    axis.text.y = element_blank(), 
    axis.text.x = element_blank(),
    axis.title.x = element_blank(),
    axis.title.y = element_blank(),
    axis.ticks = element_blank(), 
    panel.grid = element_blank()
  )




ggsave(
  filename = here::here("figures/2024-07-01_expert-panel-meeting/mitigationOROTypeValidation_Sankey.pdf"),
  plot = oroTypeValidationSankey,
  height = 7, width = 7, units="in"
)

# clean environment
rm(tempTab, temp, label_mat, label_df_long)

``` 


For revision, can we use the development stage across time? E.g. compare the number of design studies after a technology has been implemented at scale – indicates that the technology is constantly being improved. Maybe show as a stacked barplot with year on the x axis (with a line overlaid showing overall number of articles, or some metric of deployment numbers). 


```{r development stage and policy mentions over time}
df_long_list <- list()
variables <- c("ORO type", "Concrete policy mentioned?", "Development stage")
for(v in 1:length(variables)){
  # variable column of interest
  var <- unlist(scope_df[,variables[v]])
  # Get a vector of all the unique labels
  labels <- unique(var)
  labels <- labels[!is.na(labels)]
  labels <- unlist(strsplit(labels, "[|||]"))
  labels <- labels[labels != ""]
  labels <- unique(labels)
  labels <- sort(labels)
  # For each article, identify if it has been coded yes for a given label
  label_mat <- matrix(nrow = nrow(scope_df), ncol = length(labels), dimnames = list(c(scope_df$`Article ID`), labels))
  for(i in 1:nrow(scope_df)){
    for(l in 1:length(labels)){
      label_mat[i,l] <- sum(grepl(labels[l], scope_df[i,variables[v]]))
    }
  }
  label_df <- as.data.frame(label_mat)
  label_df$`Article ID` <- scope_df$`Article ID`
  # Format into long form
  label_df_long <- reshape2::melt(label_df, id.vars = c("Article ID"), 
                                variable.name = variables[v], value.name = "value")
  label_df_long <- label_df_long %>%
    filter(value == 1)
  label_df_long$value <- NULL
  label_df_long <- label_df_long %>% 
    inner_join(analysis_id_lookup, by = "Article ID") %>%
    inner_join(sampleDf %>% select(analysis_id, year), by = "analysis_id") %>%
    select(-c(`Article ID`))
  # save results
  df_long_list[[v]] <- label_df_long
  # Clean up environment from temporary variables
  if(v==length(variables)){
    rm(labels, var, label_mat, label_df, label_df_long)
  }
  
}
names(df_long_list) <- variables


# Dataframe crossing ORO type and devlopment stage to take annual sums
oroType_devStage <- df_long_list$`ORO type` %>%
  inner_join(df_long_list$`Development stage`, by = c("analysis_id","year")) %>%
  group_by(`ORO type`, `Development stage`, year) %>%
  summarise(n=n()) %>%
  mutate(year = as.Date(paste0(year,"-01-01")),
         `Development stage` = factor(`Development stage`,
                                      levels = c("Design/experiment", "Deployed - pilot", "Deployed - at scale","Modelling future placement/upscaling"))) %>%
  filter(!is.na(year))

oroType_policy <- df_long_list$`ORO type` %>%
  inner_join(df_long_list$`Concrete policy mentioned?`, by = c("analysis_id","year")) %>% 
  filter(`Concrete policy mentioned?` == "true") %>%
  group_by(`ORO type`, year) %>%
  summarise(n=n()) %>%
  mutate(year = as.Date(paste0(year,"-01-01"))) %>%
  filter(!is.na(year)) %>%
  filter(n>0)


## Plot
devStage_ggp <- ggplot(data = oroType_devStage, aes(x = year, y=n, fill = `Development stage`))+
  geom_col(width = 250)+
  geom_point(data = oroType_policy, aes(x = year, y=n), col = "deeppink", inherit.aes = FALSE)+
  geom_line(data = oroType_policy, aes(x = year, y=n), col = "deeppink", inherit.aes = FALSE, linewidth = 0.8, alpha = 0.75)+
  scale_x_date(limits = as.Date(c("2000-01-01","2022-12-31")))+
  scale_fill_viridis_d()+
  facet_wrap(vars(`ORO type`), labeller = label_wrap_gen(), drop=TRUE, scales = "free")+
  labs(y = "Number of articles",
       caption = "dots indicate number of articles mentioning concrete policy")+
  theme_bw()+
  theme(
    legend.position = "bottom"
  )


ggsave(
  filename = here::here("figures/2024-07-01_expert-panel-meeting/developmentStagePolicyAnnual_barplot.pdf"),
  plot = devStage_ggp,
  height = 7, width = 7, units="in"
)

devStage_ggp2 <- ggplot(data = oroType_devStage, aes(x = year, y=n, fill = `ORO type`))+
  geom_col(position = 'stack', colour = "darkgrey")+
  scale_x_date(limits = as.Date(c("2000-01-01","2022-12-31")))+
  scale_fill_manual(values = c(RColorBrewer::brewer.pal(5, "Blues"),
                               RColorBrewer::brewer.pal(3, "Greens")[1:2],
                               RColorBrewer::brewer.pal(3, "Reds")[1:2]))+
  facet_grid(`Development stage` ~., scales = "free_y", labeller = label_wrap_gen())+
  labs(y = "Number of articles")+
  theme_bw()
devStage_ggp2

ggsave(
  filename = here::here("figures/2024-07-01_expert-panel-meeting/developmentStageOROTypeAnnual_barplot.pdf"),
  plot = devStage_ggp2,
  height = 7, width = 7, units="in"
)



devStage_ggp3 <- ggplot(data = oroType_devStage %>%
                          group_by(`ORO type`, year) %>%
                          mutate(proportion = n/sum(n, na.rm=T)), 
                        aes(x = year, y=proportion, fill = `Development stage`))+
  geom_col(width = 250)+
  geom_point(data = oroType_policy %>%
               group_by(`ORO type`) %>%
               mutate(n_scale = scales::rescale(n, c(0,1))), 
             aes(x = year, y=n_scale), col = "deeppink", inherit.aes = FALSE)+
  geom_line(data = oroType_policy %>%
               group_by(`ORO type`) %>%
               mutate(n_scale = scales::rescale(n, c(0,1))), 
            aes(x = year, y=n_scale), col = "deeppink", inherit.aes = FALSE, linewidth = 0.8, alpha = 0.75)+
  scale_x_date(limits = as.Date(c("2000-01-01","2022-12-31")))+
  scale_fill_viridis_d()+
  facet_wrap(vars(`ORO type`), labeller = label_wrap_gen(), drop=TRUE, scales = "free")+
  labs(y = "Proportion articles",
       caption = "dots indicate number of articles mentioning concrete policy")+
  theme_bw()+
  theme(
    legend.position = "bottom"
  )


ggsave(
  filename = here::here("figures/2024-07-01_expert-panel-meeting/developmentStagePolicyAnnualProportion_barplot.pdf"),
  plot = devStage_ggp3,
  height = 7, width = 7, units="in"
)
```


```{r ORO type vs outcomes over time}
df_long_list <- list()
variables <- c("ORO type", "Outcome - effectiveness", "Outcome - co-benefit or dis-benefit")
for(v in 1:length(variables)){
  # variable column of interest
  var <- unlist(scope_df[,variables[v]])
  # Get a vector of all the unique labels
  labels <- unique(var)
  labels <- labels[!is.na(labels)]
  labels <- unlist(strsplit(labels, "[|||]"))
  labels <- labels[labels != ""]
  labels <- unique(labels)
  labels <- sort(labels)
  # For each article, identify if it has been coded yes for a given label
  label_mat <- matrix(nrow = nrow(scope_df), ncol = length(labels), dimnames = list(c(scope_df$`Article ID`), labels))
  for(i in 1:nrow(scope_df)){
    for(l in 1:length(labels)){
      label_mat[i,l] <- sum(grepl(labels[l], scope_df[i,variables[v]]))
    }
  }
  label_df <- as.data.frame(label_mat)
  label_df$`Article ID` <- scope_df$`Article ID`
  # Format into long form
  label_df_long <- reshape2::melt(label_df, id.vars = c("Article ID"), 
                                variable.name = variables[v], value.name = "value")
  label_df_long <- label_df_long %>%
    filter(value == 1)
  label_df_long$value <- NULL
  label_df_long <- label_df_long %>% 
    inner_join(analysis_id_lookup, by = "Article ID") %>%
    inner_join(sampleDf %>% select(analysis_id, year), by = "analysis_id") %>%
    select(-c(`Article ID`))
  # save results
  df_long_list[[v]] <- label_df_long
  # Clean up environment from temporary variables
  if(v==length(variables)){
    rm(labels, var, label_mat, label_df, label_df_long)
  }
  
}
names(df_long_list) <- variables


# Dataframe crossing ORO type and devlopment stage to take annual sums
oroType_effectiveness <- df_long_list$`ORO type` %>%
  inner_join(df_long_list$`Outcome - effectiveness`, by = c("analysis_id","year")) %>%
  group_by(`ORO type`, `Outcome - effectiveness`, year) %>%
  summarise(n=n()) %>%
  mutate(year = as.Date(paste0(year,"-01-01"))) %>%
  filter(!is.na(year))

oroType_coben <- df_long_list$`ORO type` %>%
  inner_join(df_long_list$`Outcome - co-benefit or dis-benefit`, by = c("analysis_id","year")) %>% 
  filter(`Outcome - co-benefit or dis-benefit` == "true") %>%
  group_by(`ORO type`, year) %>%
  summarise(n=n()) %>%
  mutate(year = as.Date(paste0(year,"-01-01"))) %>%
  filter(!is.na(year)) 


## Plot
outcome_ggp <- ggplot(data = oroType_effectiveness, aes(x = year, y=n, fill = `Outcome - effectiveness`))+
  geom_col(width = 250)+
  geom_point(data = oroType_coben, aes(x = year, y=n), col = "deeppink", inherit.aes = FALSE)+
  geom_line(data = oroType_coben, aes(x = year, y=n), col = "deeppink", inherit.aes = FALSE, linewidth = 0.8, alpha = 0.75)+
  scale_x_date(limits = as.Date(c("2000-01-01","2022-12-31")))+
  scale_fill_viridis_d()+
  facet_wrap(vars(`ORO type`), labeller = label_wrap_gen(), drop=TRUE, scales = "free")+
  labs(y = "Number of articles",
       caption = "dots indicate number of articles mentioning co-benefit or dis-benefit outcomes")+
  #guides(fill = guide_legend(nrow = 2))+
  theme_bw()+
  theme(
    legend.position = "right"
  )

ggsave(
  filename = here::here("figures/2024-07-01_expert-panel-meeting/outcomesAnnual_barplot.pdf"),
  plot = outcome_ggp,
  height = 7, width = 10, units="in"
)

devStage_ggp2 <- ggplot(data = oroType_devStage, aes(x = year, y=n, fill = `ORO type`))+
  geom_col(position = 'stack', colour = "darkgrey")+
  scale_x_date(limits = as.Date(c("2000-01-01","2022-12-31")))+
  scale_fill_manual(values = c(RColorBrewer::brewer.pal(5, "Blues"),
                               RColorBrewer::brewer.pal(3, "Greens")[1:2],
                               RColorBrewer::brewer.pal(3, "Reds")[1:2]))+
  facet_grid(`Development stage` ~., scales = "free_y", labeller = label_wrap_gen())+
  labs(y = "Number of articles")+
  theme_bw()
devStage_ggp2

ggsave(
  filename = here::here("figures/2024-07-01_expert-panel-meeting/developmentStageOROTypeAnnual_barplot.pdf"),
  plot = devStage_ggp2,
  height = 7, width = 7, units="in"
)

```

# Compare trends in knowledge development from scoping to deployments from external data


```{r format knowledge development data}

df_long_list <- list()
variables <- c("ORO type")
for(v in 1:length(variables)){
  # variable column of interest
  var <- unlist(scope_df[,variables[v]])
  # Get a vector of all the unique labels
  labels <- unique(var)
  labels <- labels[!is.na(labels)]
  labels <- unlist(strsplit(labels, "[|||]"))
  labels <- labels[labels != ""]
  labels <- unique(labels)
  labels <- sort(labels)
  # For each article, identify if it has been coded yes for a given label
  label_mat <- matrix(nrow = nrow(scope_df), ncol = length(labels), dimnames = list(c(scope_df$`Article ID`), labels))
  for(i in 1:nrow(scope_df)){
    for(l in 1:length(labels)){
      label_mat[i,l] <- sum(grepl(labels[l], scope_df[i,variables[v]]))
    }
  }
  label_df <- as.data.frame(label_mat)
  label_df$`Article ID` <- scope_df$`Article ID`
  # Format into long form
  label_df_long <- reshape2::melt(label_df, id.vars = c("Article ID"), 
                                variable.name = variables[v], value.name = "value")
  label_df_long <- label_df_long %>%
    filter(value == 1)
  label_df_long$value <- NULL
  label_df_long <- label_df_long %>% 
    inner_join(analysis_id_lookup, by = "Article ID") %>%
    inner_join(sampleDf %>% select(analysis_id, year), by = "analysis_id") %>%
    select(-c(`Article ID`))
  # save results
  df_long_list[[v]] <- label_df_long
  # Clean up environment from temporary variables
  if(v==length(variables)){
    rm(labels, var, label_mat, label_df, label_df_long)
  }
  
}
names(df_long_list) <- variables

# Take the annual sums for each oro type
annual_knowledge_dev <- df_long_list$`ORO type` %>%
  group_by(`ORO type`, year) %>%
  summarise(n_knowledge = n()) %>%
  mutate(year = as.Date(paste0(year, "-01-01"))) %>%
  filter(!is.na(year)) %>%
  mutate(ORO_type_scope = factor(`ORO type`, levels = scopeOROFactorLevels)) # %>%
  # mutate(ORO_type = recode(`ORO type`,
  #                          `MRE - marine energy` = "MRE",
  #                          `MRE - located` = "MRE",
  #                          `Efficiency - ships` = "Efficiency",
  #                          `Efficiency - fish` = "Efficiency",
  #                          `CDR - BC` = "CDR",
  #                          `CDR - CCS` = "CDR",
  #                          ))

summary(annual_knowledge_dev)

```

```{r format deployment data}

## MRE - annual by technology
df <- readxl::read_excel(here::here("data/raw-data/external/IRENA-electricity-statistics-by-country-year.xlsx"),
                         sheet = "ELECSTAT-C", col_names = FALSE, skip = 2)
colnames(df) <- c("country","technology","data_type","grid_connection","year","value")

df <- df %>% 
  tidyr::fill(country, technology, data_type, data_type, grid_connection, .direction = "down")

# filter data to one variable, all grid connections
df <- df %>%
  filter(data_type == "Electricity Installed Capacity (MW)", grid_connection == "All", technology %in% c("Marine energy","Offshore wind energy")) 

# Format columns
df$year <- as.Date(paste0(df$year,"-01-01"))

# Summarise annual counts
mre_df <- df %>%
  group_by(year, technology) %>%
  rename(group = technology) %>%
  summarise(deployment_count = sum(value, na.rm=T)) %>%
  mutate(dataset = "IRENA installed capacity (MW)",
         ORO_type_scope = ifelse(grepl("Marine energy", group), "MRE - marine energy","MRE - located")) %>%
  select(year, ORO_type_scope, dataset, deployment_count)



## Efficiency -- annual by shipping and fishing

# IEA domestic freight transport energy efficiency (gCO2/tkm)
# https://www.iea.org/data-and-statistics/data-product/energy-efficiency-indicators
domestic_shipping_efficiency <- readxl::read_excel(here::here("data/raw-data/external/IEA-shipping-energy-efficiency.xlsx"), sheet = "Data")
domestic_shipping_efficiency$year <- as.Date(paste0(domestic_shipping_efficiency$year,"-01-01"))

# Fishing Emissions intensity - Greer et al 2019
# For extraction code using metadigitize see version of file 0.2.1
load(here::here("data/raw-data/external/Greer-et-al-2019_emissions-intensity_industrial.RData"))

eff_df <- rbind(
  domestic_shipping_efficiency %>% 
    mutate(deployment_count = 1/carbon_intensity_gCO2_per_tkm) %>% 
    mutate(ORO_type_scope = "Efficiency - ships",
           dataset = "IEA domestic freight transport energy efficiency (tkm/gCO2)") %>% 
    select(year, ORO_type_scope, dataset, deployment_count),
  greer_industrial %>%
    mutate(
      year = as.Date(paste0(round(x),"-01-01")),
      ORO_type_scope = "Efficiency - fish",
      dataset = "Greer et al. (2019) Fishing efficiency (tcatch/tCO2)",
      deployment_count = 1/y
    ) %>%
    select(year, ORO_type_scope, dataset, deployment_count)
)



##  CDR 
# Blue carbon restoration  
bc_restoration <- readxl::read_excel(here::here("data/raw-data/external/Duarte2020RestorationDataSet-1.xlsx"),
                                     sheet = "Data")
bc_restoration$year <- as.Date(paste0(bc_restoration$Year, "-01-01"))
bc_restoration <- bc_restoration %>% filter(as.Date("1990-01-01") <= year)

bc_restoration <- bc_restoration %>%
  group_by(year) %>%
  summarise(deployment_count = n()) %>%
  mutate(
    ORO_type_scope = "CDR - BC",
    dataset = "BC restoration projects (Duarte et al 2020)"
  ) %>%
  select(year, ORO_type_scope, dataset, deployment_count)


# Patents
patents_df <- read.csv(here::here("data/raw-data/external/State-of-Carbon-Report-2024_Chatper-2_Number_of_patents.csv"))
colnames(patents_df)[6:28] <- gsub("X","",colnames(patents_df)[6:28])
patents_df <- patents_df %>%
  select(-c(Model, Scenario, Region, Unit))
patents_df <- patents_df %>%
  filter(Variable %in% c("Research|Removal|Patents|Coastal Wetland Restoration", "Research|Removal|Patents|Ocean Alkalinity Enhancement","Research|Removal|Patents|Ocean Fertilization"))
patents_df <- reshape2::melt(patents_df, 
                                   id.vars = c("Variable"), 
                                   variable.name = "year", 
                                   value.name = "value")
patents_df$year <- as.Date(paste0(patents_df$year,"-01-01"))
patents_df <- patents_df %>%
  mutate(ORO_type_scope = as.character(factor(Variable, 
                           levels = c("Research|Removal|Patents|Coastal Wetland Restoration", "Research|Removal|Patents|Ocean Alkalinity Enhancement","Research|Removal|Patents|Ocean Fertilization"),
                           labels = c("CDR - BC", "CDR - OAE", "CDR - OIF"))),
         dataset = "Number of patents",
         deployment_count = value) %>%
  select(year, ORO_type_scope, dataset, deployment_count)


# Investment in startups
startup_investments_df <- read.csv(here::here("data/raw-data/external/State-of-Carbon-Report-2024_Chatper-3_Investment_in_CDR_start-ups.csv"))
colnames(startup_investments_df)[6:20] <- gsub("X","",colnames(startup_investments_df)[6:20])
startup_investments_df <- startup_investments_df %>%
  select(-c(Model, Scenario, Region, Unit))

startup_investments_df <- startup_investments_df %>%
  filter(Variable %in% c("Start-Up|Removal|Funding|Biomass Sinking", "Start-Up|Removal|Funding|Direct Ocean Capture", "Start-Up|Removal|Funding|Ocean Alkalinity Enhancement"))

startup_investments_df <- reshape2::melt(startup_investments_df, 
                                   id.vars = c("Variable"), 
                                   variable.name = "year", 
                                   value.name = "value")
startup_investments_df$year <- as.Date(paste0(startup_investments_df$year,"-01-01"))

startup_investments_df <- startup_investments_df %>%
  mutate(ORO_type_scope = as.character(factor(Variable, 
                           levels = c("Start-Up|Removal|Funding|Biomass Sinking", "Start-Up|Removal|Funding|Direct Ocean Capture", "Start-Up|Removal|Funding|Ocean Alkalinity Enhancement"),
                           labels = c("CDR - Macroalgae","CDR - CCS", "CDR - OAE"))),
         dataset = "Start-up investment (Billion USD)",
         deployment_count = value) %>%
  select(year, ORO_type_scope, dataset, deployment_count)
  

# OIF fertilization experiments
oif_df <- readxl::read_excel(here::here("data/raw-data/external/OIF_data.xlsx"))
oif_df <- oif_df %>%
  filter(!is.na(Year)) %>%
  mutate(year = as.Date(paste0(Year,"-01-01"))) %>%
  group_by(year) %>%
  summarise(deployment_count = sum(`Total Size`, na.rm=T)) %>%
  mutate(
    ORO_type_scope = "CDR - OIF",
    dataset = "OIF fertilized area (km2)"
  ) %>%
  select(year, ORO_type_scope, dataset, deployment_count)

  


## Bind all deployment datasets together
annual_deployment_count <- rbind(
  mre_df,
  eff_df,
  patents_df,
  startup_investments_df,
  oif_df
)

annual_deployment_count$ORO_type_scope <- factor(annual_deployment_count$ORO_type_scope, 
                                                 levels = scopeOROFactorLevels)
```


```{r plot knowledge development vs deployment count}

# Scale all values between 0 and 1
annual_knowledge_dev <- annual_knowledge_dev %>%
  group_by(ORO_type_scope) %>%
  mutate(n_knowledge_scaled = scales::rescale(n_knowledge, c(0,1)))

annual_deployment_count <- annual_deployment_count %>%
  group_by(ORO_type_scope, dataset) %>%
  mutate(deployment_count_scaled = scales::rescale(deployment_count, c(0,1)))

know_deploy_ggp <- ggplot()+
  geom_col(data = annual_knowledge_dev, aes(x = year, y = n_knowledge_scaled), width = 250)+
  geom_line(data = annual_deployment_count, aes(x = year, y = deployment_count_scaled, color = dataset), linewidth = 1.2)+
  facet_wrap(vars(ORO_type_scope), scales = "free") +
  scale_x_date()+
  scale_colour_discrete(labels = label_wrap_gen())+
  labs(y = "scaled value")+
  guides(color = guide_legend(title = "Data source", nrow = 2))+
  theme_bw()+
  theme(
    legend.position = "bottom"
  )
  

ggsave(
  filename = here::here("figures/2024-07-01_expert-panel-meeting/knowledgeDev_vs_deployments.pdf"),
  plot = know_deploy_ggp,
  height = 7, width = 7, units="in"
)

```
