==> analyses/1.0_write-predictions-to-sql.Rmd <==
---
title: "0.9_format-article-answers_notMRE"
author: "Devi Veytia"
date: "2025-03-17"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


Write all tables into sql database


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r load libraries}
library(dplyr)
library(dbplyr)
library(R.utils)
library(RSQLite)
```

```{r set the seed}
addTaskCallback(function(...) {set.seed(123);TRUE})
```


```{r set sql table name and comment}
p1_db <- RSQLite::dbConnect(RSQLite::SQLite(),
                      here::here("data","sqlite-databases","product1.sqlite"),
                                 create=FALSE)

```


# Update the systematic map on OROs

```{r Update the search results table of unique references - uniquerefs}

uniquerefs_2022 <- tbl(p1_db, "uniquerefs") %>% collect()
range(uniquerefs_2022$analysis_id) # 1 392110

uniquerefs_2025 <- read.delim(
  here::here("data/raw-data/unique_references_UPDATE_13-05-2025.txt"), quote=""
)
uniquerefs_2025$year <- as.character(uniquerefs_2025$year)
uniquerefs_2025$duplicate_id <- as.character(uniquerefs_2025$duplicate_id)
range(uniquerefs_2025$analysis_id) # 392111 561557

cols <- intersect(colnames(uniquerefs_2022), colnames(uniquerefs_2025))


dbWriteTable(conn=p1_db, 
             name="uniquerefs_update2025", 
             value=uniquerefs_2022[,cols] %>% bind_rows(uniquerefs_2025[,cols]),
             append=FALSE, 
             overwrite = FALSE)

rm(uniquerefs_2022, uniquerefs_2025)
```

```{r update oro screen and branch predictions}
updatePredDir <- here::here("outputs/coded-label-predictions/ORO_sys_map_update")
updatePredFiles <- dir(updatePredDir)
varNames <- gsub("_update_predictions.csv","", updatePredFiles)
tableNames <- paste0("pred_", varNames)
tableNames[tableNames == "pred_oro_screen"] = "pred_relevance"
tableNames_update = paste0(tableNames,"_update")

existing_tables <- dbListTables(p1_db)

for(i in 1:length(tableNames)){
  if(tableNames_update[i] %in% existing_tables){
    next
  }else{
    old_table <- tbl(p1_db, tableNames[i]) %>% collect()
    update_table <- read.csv(file.path(updatePredDir, updatePredFiles[i])) %>%
      rename(analysis_id = id)
    colnames(update_table) = gsub("\\.\\.\\."," - ", colnames(update_table))
    update_table = old_table %>%
      bind_rows(update_table)
    
    
    dbWriteTable(conn=p1_db,
               name=tableNames_update[i],
               value=update_table,
               append=FALSE,
               overwrite = FALSE)
  }
  
}

rm(old_table, update_table)

```

```{r explore sample of new predictions to check}
uniquerefs <- tbl(p1_db, "uniquerefs_update2025")
pred_oro_branch <- tbl(p1_db, "pred_oro_branch_update")

newTitles <- pred_oro_branch %>%
  filter(392110 < analysis_id, 0.5 <= `oro_branch.Mitigation - mean_prediction`) %>%
  left_join(uniquerefs) %>%
  select(analysis_id, title) %>%
  collect()

nrow(newTitles) # 14841 new mitigation branch articles

View(newTitles) # check -- yes they are mitigation. The article id indexing is correct
rm(newTitles)
```


```{r disconnect databases after updating sys map predictions}
DBI::dbDisconnect(p1_db)
```


# Write predictions for Product 1 variables


```{r Write predictions for Product 1 variables}

overwriteDB = TRUE


p1_db <- RSQLite::dbConnect(RSQLite::SQLite(),
                      here::here("data","sqlite-databases","product1.sqlite"),
                                 create=FALSE)

prod1pred_files = dir(here::here("outputs/coded-label-predictions"))
prod1pred_files = prod1pred_files[
  !(prod1pred_files %in% c("ecosystem_1_predictions.csv",
                           "ORO_sys_map_update",
                           "oro_type.MRE_v2_MreLO_predictions.csv",
                           "oro_type_v1_predictions.csv"))
]



# also add seen articles
cod1 <- read_delim(here::here("data/derived-data/screen-article-answers-formatted/articleAnswers_formatted_2025-03-17_v2.txt"))
cod2 <- read_delim(here::here("data/derived-data/screen-article-answers-formatted/articleAnswers_notMRE_formatted_2025-05-27.txt"))
seen_df <- cod1 %>%
  bind_rows(cod2) %>%
  distinct(id, .keep_all = TRUE) %>%
  rename(analysis_id = id)

## function to format data to long
format_pred_2_long <- function(df,var){
  
  mean_long <- df %>% select(contains("mean_prediction"), analysis_id) 
  colnames(mean_long) <- gsub(paste0(var, "."),"", colnames(mean_long))
  colnames(mean_long) <- gsub("mean_prediction","",colnames(mean_long))
  colnames(mean_long) <- gsub("_$","",colnames(mean_long))
  std_long <- df %>%select(contains("std_prediction"), analysis_id)
  colnames(std_long) <- gsub(paste0(var, "."),"", colnames(std_long))
  colnames(std_long) <- gsub("std_prediction","",colnames(std_long))
  colnames(std_long) <- gsub("_$","",colnames(std_long))
    

  mean_long <- reshape2::melt(data=mean_long, id.vars = c("analysis_id"), variable.name = "level", value.name = "mean_prediction")
  std_long <- reshape2::melt(data = std_long, id.vars = c("analysis_id"), variable.name = "level", value.name = "std_prediction")
  df_long <- merge(mean_long, std_long, all = FALSE)
  df_long <- df_long %>%
    filter(0.5 <= mean_prediction)
  
  return(df_long)

}


## Loop through files and save to sql
for(f in prod1pred_files){
  df = read.csv(here::here("outputs/coded-label-predictions",f))
  colnames(df) <- gsub("\\.\\.\\.","_", colnames(df))
  df <- df %>% rename(analysis_id = id)
  
  
  
  if(grepl("ecosystem|oro_type", f)){
    
    var <- gsub("[.].*","",colnames(df)[-1])
    var <- var[1]
    
    # add in the seen data 
    seen_tmp <- seen_df %>%
      filter(!(analysis_id %in% df$analysis_id)) %>%
      select(c(analysis_id, contains(var)))
    colnames(seen_tmp) <- gsub(" ",".", colnames(seen_tmp))
    colnames(seen_tmp)[-1] <- paste0(colnames(seen_tmp)[-1], "_mean_prediction")
    df <- df %>%
      bind_rows(seen_tmp) %>%
      distinct(analysis_id, .keep_all = TRUE) %>%
      filter(!is.na(analysis_id))
    
    
    df_long <- format_pred_2_long(df, var)
    
    # colnames(df) <- gsub(paste0(var, "."),"", colnames(df))
    if(var == "oro_type"){var <- "oro_type_mit"}
    df$variable <- paste(var)
    df_long$variable <- paste(var)
    
    dbWriteTable(conn=p1_db, 
             name=paste("pred",var,"long", sep="_"), 
             value=df_long,
             append=FALSE, 
             overwrite = overwriteDB)
    
    dbWriteTable(conn=p1_db, 
             name=paste("pred",var, sep="_"), 
             value=df,
             append=FALSE, 
             overwrite = overwriteDB)
    
    
    
  }else{
    var <- gsub("_mean_prediction","",colnames(df)[grep("_mean_prediction", colnames(df))])
    colnames(df) <- gsub(paste0(var, "_"),"", colnames(df))
    df$variable <- paste(var)
    dbWriteTable(conn=p1_db, 
             name= paste("pred",var, sep="_"),
             value=df,
             append=FALSE, 
             overwrite = overwriteDB)
  }
  
  
  
}

DBI::dbDisconnect(p1_db)

# tables added: pred_ecosystem, pred_ecosystem_long, pred_oro_type_mit, pred_oro_type_mit_long, pred_outcome_effectiveness, pred_outcome_quantitative, pred_primary_research

```
