==> analyses/2.0_compile-time-series-data.Rmd <==
---
title: "2.0_compile-trends-data"
author: "Devi Veytia"
date: "2025-07-08"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Note this version is different from the data compilation in 0.1_2025_OOSC_figures.Rmd because I had to remove CDR-Cult as this did not predict well for ORO type. Also the ORO type N publications reflect the final version of the model.


```{r load libraries}

# general data handing
library(dplyr)
library(dbplyr)
library(R.utils)
library(ggplot2)
library(ggalluvial)
library(tidyr)
library(stringr)
library(viridis)
library(countrycode)
library(broom)
library(conflicted)
library(tidyverse)


conflict_prefer("select", "dplyr")
conflicts_prefer(dplyr::filter)



## AESTHETICS
factor_aes <- readxl::read_excel(here::here("R/mitigation_factor_aesthetics.xlsx"))
typeAES <- factor_aes[which(factor_aes$variable == "oro_type"),]
typeAES <- typeAES[order(typeAES$order),]

continent_levels <- c("Americas","Asia","Europe","Oceania","Africa")

## For parsing countries
countries_ls <- data.frame(name_en = countrycode::codelist$country.name.en) %>%
  mutate(country_iso = countrycode::countrycode(sourcevar   = name_en,
                                       origin      = "country.name",
                                       destination = "iso3c"),
         country=name_en)



source(here::here("R", "functions_to_format.R"))

```





Data structure:
list with each level is 
- data frame for publications: 
- data frame for n policy documents
- data frame for deployment

id variables: oro_type, component (publication, policy, deployment), variable_name,
response variable: y

## Number of publications

```{r read in oro type predictions}

# get id and year metadata
require(RSQLite)
p1_db <- RSQLite::dbConnect(RSQLite::SQLite(),
                      here::here("data","sqlite-databases","product1.sqlite"),
                                 create=FALSE)

uniquerefs <- tbl(p1_db, "uniquerefs_update2025") %>%
  select(analysis_id, year) 

predOroType <- tbl(p1_db, "pred_oro_type_mit_long") %>%
  left_join(uniquerefs, by = "analysis_id") %>%
  collect()

dbDisconnect(p1_db)

# read in predictions and join with year


# how many relevant for each oro type?
pubs <- predOroType %>%
  filter(!is.na(year), 0.5 <= mean_prediction) %>%
  group_by(level, year) %>%
  summarise(
    y = n_distinct(analysis_id)
  )%>%
  mutate(
    component = "publications",
    variable_name = "Publications (N)",
    level = gsub("[.]","-", level),
    year= as.numeric(year)
  ) %>%
  rename(oro_type = level)

pubs_MRE_gen <- pubs %>%
  filter(grepl("MRE", oro_type)) %>%
  group_by(year, component, variable_name) %>%
  summarise(
    y = sum(y, na.rm=T)
  )%>%
  mutate(
    oro_type = "MRE-General"
  )

pubs_CDR_gen <- pubs %>%
  filter(grepl("CDR", oro_type)) %>%
  group_by(year, component, variable_name) %>%
  summarise(
    y = sum(y, na.rm=T)
  )%>%
  mutate(
    oro_type = "CDR-General"
  )

pubs <- pubs %>%
  bind_rows(pubs_MRE_gen) %>%
  bind_rows(pubs_CDR_gen)


ggplot(pubs, aes(x=year, y=y, alluvium=oro_type, fill=oro_type, colour=oro_type))+
  geom_alluvium(alpha = 0.8, decreasing=FALSE)+
  theme_minimal()


## All OROs together
all_Pub <- predOroType %>%
  filter(!is.na(year), 0.5 <= mean_prediction) %>%
  group_by(year) %>%
  summarise(
    y = n_distinct(analysis_id)
  )%>%
  mutate(
    component = "publications",
    variable_name = "Publications (N)",
    oro_type = "All mitigation OROs",
    year= as.numeric(year)
  ) %>%
  select(all_of(colnames(pubs)))


## Save
save(pubs, file = here::here("data", "derived-data", "mitigationORO_pubs.RData"))
save(all_Pub, file = here::here("data", "derived-data", "mitigationORO_pubs_allMitigationOROs.RData"))
```

```{r publications by country}

# get id and year metadata
require(RSQLite)
p1_db <- RSQLite::dbConnect(RSQLite::SQLite(),
                      here::here("data","sqlite-databases","product1.sqlite"),
                                 create=FALSE)

uniquerefs <- tbl(p1_db, "uniquerefs_update2025") %>%
  select(analysis_id, year, affiliation) 

predOroTypeCountry <- tbl(p1_db, "pred_oro_type_mit_long") %>%
  left_join(uniquerefs, by = "analysis_id") %>%
  collect()

dbDisconnect(p1_db)

## Parse countries from author affiliations

predOroTypeCountry <- predOroTypeCountry %>%
  filter(!is.na(year), 0.5 <= mean_prediction)
  
# Extract countries from affiliations  
predOroTypeCountry$countries <- extract_all_affiliation(
  predOroTypeCountry$affiliation, countries_ls)

# explode data frame so that its long form
predOroTypeCountry <- predOroTypeCountry %>%
  mutate(
    country = str_split(countries, ",\\s*"),
    country = map(country, ~ unique(str_trim(str_replace_all(.x, "^['\"]|['\"]$", ""))))
  ) %>%
  unnest(country) %>%
  filter(!is.na(country), country != "") %>%
  # add weighting 
  group_by(analysis_id) %>%
  mutate(
    country_weight = 1/n(),
    country_count = 1
  )

# how many relevant for each oro type x country?
pubs_country <- predOroTypeCountry %>%
  filter(!is.na(year), 0.5 <= mean_prediction) %>%
  group_by(level, year, country) %>%
  summarise(
    y = sum(country_weight, na.rm=T)
  )%>%
  mutate(
    component = "publications",
    variable_name = "Publications (N)",
    level = gsub("[.]","-", level),
    year= as.numeric(year)
  ) %>%
  rename(oro_type = level)



pubs_country$continent <- countrycode(sourcevar = pubs_country$country,
                            origin = "country.name",
                            destination = "continent")

pubs_country <- pubs_country %>%
  # filter(
  #   continent %in% continent_levels
  # )%>%
  mutate(
    country_simple = case_when(
      grepl("United States", country) ~ country,
      grepl("China", country) ~ country,
      grepl("United Kingdom", country) ~ country,
      grepl("Europe", continent) ~ continent,
      TRUE ~ "Other"
    ),
    continent = factor(continent, levels = continent_levels),
    country_simple = factor(
      country_simple, 
      levels = c("United Kingdom","Europe","United States","China","Other")
    )
  )


## quick plot
pubs_country %>%
  group_by(country_simple, year)%>%
  summarise(
    y=sum(y)
  )%>%
  ggplot()+
  geom_area(aes(x=year, y= y, fill=country_simple), position="stack")+
  theme_minimal()


## Save
save(pubs_country, file = here::here("data", "derived-data", "mitigationORO_pubs_country.RData"))

```


```{r Load in number of O&C publications to calculate proportion?}
## Load in tabulated number of publications from previous chunk
load(here::here("data", "derived-data", "mitigationORO_pubs.RData"))


## Join with the number of O&C publications to get the proportion of ORO publications 
## NOTE I need to make a 2025 update
wosFile <- here::here("data/external/ocean-and-climate-publications/filename.txt")

# get OandA publications (worldwide)
tempOA <- read.delim(wosFile) %>%
rename(
  year = Publication.Years,
  n_OC = Record.Count
) %>%
select(year, n_OC) %>%
mutate(year = as.numeric(year))



# Join together and calculate proportion
pubsProp <- pubs %>%
    left_join(tempOA, by = "year") %>%
    mutate(y = y/n_OC,
           variable_name = "Publications (prop)"
           ) %>%
    filter(!is.na(year)&
           !is.na(n_ORO) &
           !is.na(y)) %>%



## Do I calculate the weighted sum or count one document multiple times? For now comment out weighted sum calculations
# docTotals <- table(predMitFr$analysis_id) %>% as.data.frame(responseName = "doc_weight") 
# colnames(docTotals) <- c("analysis_id","doc_weight")
# docTotals[,2] <- 1/docTotals[,2]
# docTotals$analysis_id <- as.double(as.character(docTotals$analysis_id))



## Format for consistency with id variables:
# id variables: oro_branch, oro_type, country_name, country_iso, component (publication, policy, deployment), variable_name,
# response variable: y

# pubsProp <- pubsProp %>%
#   mutate(
#         component = "publications",
#         variable_name = "Publications (prop)"
#   ) %>%
#   reshape2::melt(id.vars = c("country_name","country_iso","component","oro_branch","oro_type","year"),
#        value.name = "y", variable.name = "variable_name")




# save(pubsProp, file = here::here("data", "derived-data", "mitigationORO_pubsProp.RData"))
```


## N policy documents

```{r read in legislation documents}

## Read in and format policy matches

all_legDat <- read.csv(here::here("data/derived-data/en-fr-de-sp_fullTextQueryMatches_allOROs_long.csv")) %>%
  rename(country_name = Country_short, year=Year, oro_type = oroType) %>%
  filter(1975 < year,
         Binding.legislation == "True", 
         National.level == "True"
         ) %>%
  mutate(
    oro_type = case_when(
      oro_type == "M_MRE_general" ~ "MRE-General", ## Note to fix this later
      oro_type == "M_MRE_ocean" ~ "MRE-Ocean",
      oro_type == "M_MRE_located" ~ "MRE-Located",
      oro_type == "M_MRE_bio" ~ "MRE-Bio",
      oro_type == "M_Incr_eff" ~ "Efficiency",
      oro_type == "M_CCS" ~ "CCS",
      oro_type == "M_CDR_other" ~ "CDR-General",
      oro_type == "M_CDR_bc" ~ "CDR-BC",
      oro_type == "M_CDR_oae" ~ "CDR-OAE",
      oro_type == "M_CDR_biopump" ~ "CDR-BioPump",
      # oro_type == "M_CDR_cult" ~ "CDR-Cult",
     
      TRUE ~ "Other"
    )) %>%
  filter(oro_type != "Other") 


# Tabulate all documents
legDat <- all_legDat %>%
  group_by(oro_type, year) %>%
  summarise(y = n_distinct(search_id)) %>%
  mutate(
    component = "legislation",
    variable_name = "Legislation (N)"
  )

all_legDat <- all_legDat %>%
  group_by(year) %>%
  summarise(y = n_distinct(search_id)) %>%
  mutate(
    oro_type = "All mitigation OROs",
    component = "legislation",
    variable_name = "Legislation (N)"
  )


# # For 'MRE', add to both Located and Ocean
# MRE_sub <- legDat[legDat$oro_type == "MRE",]
# MRE_sub <- MRE_sub %>% rename(y_MRE = y) %>% select(y_MRE, year)
# legDat <- legDat[legDat$oro_type != "MRE",]
# legDat <- legDat %>%
#   left_join(MRE_sub, by="year") %>% 
#   replace_na(list(y_MRE = 0)) %>%
#   mutate(y=ifelse(oro_type %in% c("MRE-Located","MRE-Ocean"), y_MRE + y, y))
# rm(MRE_sub)
# legDat$y_MRE <- NULL

## quick plot
  
pol_stackedBP <- ggplot(legDat, aes(x=year, y=y, fill=oro_type))+
  geom_col(position="stack")+
  # scale_fill_manual(name = "ORO type",values = typeAES$colour, breaks = typeAES$level)+
  theme_minimal()+
  theme(legend.position = "bottom")
pol_stackedBP

## save
save(legDat, file = here::here("data", "derived-data", "n_legislation_docs.RData"))
save(all_legDat, file = here::here("data", "derived-data", "n_legislation_docs_allMitigationOROs.RData"))
```

```{r read in non-binding policy documents}

## Read in and format policy matches

all_polDat <- read.csv(here::here("data/derived-data/en-fr-de-sp_fullTextQueryMatches_allOROs_long.csv")) %>%
  rename(country_name = Country_short, year=Year, oro_type = oroType) %>%
  filter(1975 < year,
         Binding.legislation == "False", 
         National.level == "True"
         ) %>%
  mutate(
    oro_type = case_when(
      oro_type == "M_MRE_general" ~ "MRE-General", ## Note to fix this later
      oro_type == "M_MRE_ocean" ~ "MRE-Ocean",
      oro_type == "M_MRE_located" ~ "MRE-Located",
      oro_type == "M_MRE_bio" ~ "MRE-Bio",
      oro_type == "M_Incr_eff" ~ "Efficiency",
      oro_type == "M_CCS" ~ "CCS",
      oro_type == "M_CDR_other" ~ "CDR-General",
      oro_type == "M_CDR_bc" ~ "CDR-BC",
      oro_type == "M_CDR_oae" ~ "CDR-OAE",
      oro_type == "M_CDR_biopump" ~ "CDR-BioPump",
      # oro_type == "M_CDR_cult" ~ "CDR-Cult",
     
      TRUE ~ "Other"
    )) %>%
  filter(oro_type != "Other")

# Tabulate all documents
polDat <- all_polDat %>%
  group_by(oro_type, year) %>%
  summarise(y = n_distinct(search_id)) %>%
  mutate(
    component = "non-binding policy",
    variable_name = "Non-binding policy (N)"
  )

all_polDat <- all_polDat %>%
  group_by(year) %>%
  summarise(y = n_distinct(search_id)) %>%
  mutate(
    oro_type = "All mitigation OROs",
    component = "non-binding policy",
    variable_name = "Non-binding policy (N)"
  )

# # For 'MRE', add to both Located and Ocean
# MRE_sub <- polDat[polDat$oro_type == "MRE",]
# MRE_sub <- MRE_sub %>% rename(y_MRE = y) %>% select(y_MRE, year)
# MRE_sub$oro_type <- NULL
# polDat <- polDat[polDat$oro_type != "MRE",]
# polDat <- polDat %>%
#   left_join(MRE_sub, by="year") %>% 
#   replace_na(list(y_MRE = 0)) %>%
#   mutate(y=ifelse(oro_type %in% c("MRE-Located","MRE-Ocean"), y_MRE + y, y))
# rm(MRE_sub)
# polDat$y_MRE <- NULL

## quick plot
  
pol_stackedBP <- ggplot(polDat, aes(x=year, y=y, fill=oro_type))+
  geom_col(position="stack")+
  # scale_fill_manual(name = "ORO type",values = typeAES$colour, breaks = typeAES$level)+
  theme_minimal()+
  theme(legend.position = "bottom")
pol_stackedBP

## save
save(polDat, file = here::here("data", "derived-data", "n_nonBindPolicy_docs.RData"))
save(all_polDat, file = here::here("data", "derived-data", "n_nonBindPolicy_docs_allMitigationOROs.RData"))
```

```{r legislation and policy by country}

leg_pol_country <- read.csv(here::here("data/derived-data/en-fr-de-sp_fullTextQueryMatches_allOROs_long.csv")) %>%
  rename(country = Country_short, year=Year, oro_type = oroType) %>%
  filter(1975 < year,
         National.level == "True"
         ) %>%
  mutate(
    oro_type = case_when(
      oro_type == "M_MRE_general" ~ "MRE-General", ## Note to fix this later
      oro_type == "M_MRE_ocean" ~ "MRE-Ocean",
      oro_type == "M_MRE_located" ~ "MRE-Located",
      oro_type == "M_MRE_bio" ~ "MRE-Bio",
      oro_type == "M_Incr_eff" ~ "Efficiency",
      oro_type == "M_CCS" ~ "CCS",
      oro_type == "M_CDR_other" ~ "CDR-General",
      oro_type == "M_CDR_bc" ~ "CDR-BC",
      oro_type == "M_CDR_oae" ~ "CDR-OAE",
      oro_type == "M_CDR_biopump" ~ "CDR-BioPump",
      # oro_type == "M_CDR_cult" ~ "CDR-Cult",
     
      TRUE ~ "Other"
    ),
    component = ifelse(
      Binding.legislation == "True", "legislation","non-binding policy"),
    variable_name = ifelse(
      Binding.legislation=="True", "Legislation (N)","Non-binding policy (N)")
    ) %>%
  filter(oro_type != "Other") %>%
  # add weighting 
  group_by(search_id, component) %>%
  mutate(
    country_weight = 1/n(),
    country_count = 1
  ) %>%
  # tabulate -- how many relevant for each oro type x country?
  group_by(oro_type, year, country, component, variable_name) %>%
  summarise(
    y = sum(country_weight)
  ) 




leg_pol_country$continent <- countrycode(sourcevar = leg_pol_country$country,
                            origin = "country.name",
                            destination = "continent")
leg_pol_country$continent[leg_pol_country$country=="European Union"] <- "Europe"
leg_pol_country$continent[is.na(leg_pol_country$continent)] <- "Not found"


leg_pol_country <- leg_pol_country %>%
  mutate(
    country_simple = case_when(
      grepl("United States", country, ignore.case=T) ~ country,
      grepl("China", country, ignore.case=T) ~ country,
      grepl("United Kingdom", country, ignore.case=T) ~ country,
      grepl("Europe", continent, ignore.case=T) ~ continent,
      TRUE ~ "Other"
    ),
    continent = factor(continent, levels = c(continent_levels,"Not found")),
    country_simple = factor(
      country_simple, 
      levels = c("United Kingdom","Europe","United States","China","Other")
    )
  )


## quick plot
leg_pol_country %>%
  group_by(country_simple, year, component)%>%
  summarise(
    y=sum(y)
  )%>%
  ggplot()+
  facet_wrap(vars(component))+
  geom_area(aes(x=year, y= y, fill=country_simple), position="stack")+
  theme_minimal()

leg_pol_country %>%
  group_by(continent, year)%>%
  summarise(
    y=sum(y)
  )%>%
  ggplot()+
  geom_area(aes(x=year, y= y, fill=continent), position="stack")+
  theme_minimal()


## Save
save(leg_pol_country, file = here::here("data", "derived-data", "n_policy_leg_country.RData"))

```


## N deployments




Difficult to get a dataset with enough longevity, so could combine indicators?

N patents from altmetrics (long, country and year available)


field trials (time coverage ~ 2018 - 2023, spatially resolved)
- (Ocean visions field trail database)[]

mCDR companies/ start-ups : goal: number of startups by founding year (recent. Need to get country and year)
-  https://community.oceanvisions.org/organizations - company names
- https://www.cdr.fyi/leaderboards - company names, tons sold, tons delivered
- GESAMP The current proposals on ocean climate intervention that GESAMP is aware of as of 31st May 2024: GESAMP_wg41_ocean_climate_intervention_projects_31_may_2024.xlsx http://www.gesamp.org/work/groups/41 ~ 184 company names
- OceanNETs_D18_oceanbased_CDR_companies ~ 50 companies with country by no year


I think for now, stick with patents and field trials, and maybe OceanNET CDR companies



```{r MRE deployment - Installed capacity}
## load data
# Data source: IRENA renewable electricity statistics (https://www.irena.org/Publications/2024/Jul/Renewable-energy-statistics-2024)

MreDat <- readxl::read_excel(here::here("data/raw-data/external/IRENA-electricity-statistics-by-country-year-2024.xlsx"),
                         sheet = "Country") %>% 
  filter(
    Technology %in% c("Marine energy","Offshore wind energy")
  ) %>% 
  mutate(
    oro_type = case_when(
      Technology == "Marine energy" ~ "MRE-Ocean",
      Technology == "Offshore wind energy" ~ "MRE-Located"
    )
  ) %>%
  rename(year = Year)  %>%
  group_by(year,oro_type) %>% 
  summarise(y = sum(`Electricity Installed Capacity (MW)`, na.rm=T)) %>%
  # add identification columns
  mutate(
    year = as.numeric(year),
    component = "deployment",
    variable_name = "Electricity Installed Capacity (MW)"
    
  )




MreDat_gen <- MreDat %>%
  group_by(year, component, variable_name) %>%
  summarise(
    y = sum(y, na.rm=T)
  )%>%
  mutate(
    oro_type = "MRE-General"
  )

MreDat <- MreDat %>%
  bind_rows(MreDat_gen)

ggplot(MreDat, aes(x=year, y=y, fill=oro_type))+
  geom_col(position="stack")+
  # scale_fill_manual(name = "ORO type",values = typeAES$colour, breaks = typeAES$level)+
  theme_minimal()+
  theme(legend.position = "bottom")


```



```{r MRE deployment - Installed capacity - by country}
## load data
# Data source: IRENA renewable electricity statistics (https://www.irena.org/Publications/2024/Jul/Renewable-energy-statistics-2024)

MreDat_country <- readxl::read_excel(here::here("data/raw-data/external/IRENA-electricity-statistics-by-country-year-2024.xlsx"),
                         sheet = "Country") %>% 
  filter(
    Technology %in% c("Marine energy","Offshore wind energy")
  ) %>% 
  mutate(
    oro_type = case_when(
      Technology == "Marine energy" ~ "MRE-Ocean",
      Technology == "Offshore wind energy" ~ "MRE-Located"
    )
  ) %>%
  rename(year = Year, country = Country)  %>%
  group_by(year,oro_type, country) %>% 
  summarise(y = sum(`Electricity Installed Capacity (MW)`, na.rm=T)) %>%
  # add identification columns
  mutate(
    year = as.numeric(year),
    component = "deployment",
    variable_name = "Electricity Installed Capacity (MW)"
    
  )


# Parse country names
MreDat_country$country <- countrycode(sourcevar = MreDat_country$country,
                            origin = "country.name",
                            destination = "country.name")

MreDat_country$continent <- countrycode(sourcevar = MreDat_country$country,
                            origin = "country.name",
                            destination = "continent")


MreDat_country$continent[is.na(MreDat_country$continent)] <- "Not found"


MreDat_country <- MreDat_country %>%
  mutate(
    country_simple = case_when(
      grepl("United States", country, ignore.case=T) ~ country,
      grepl("China", country, ignore.case=T) ~ "China",
      grepl("United Kingdom", country, ignore.case=T) ~ country,
      grepl("Europe", continent, ignore.case=T) ~ continent,
      TRUE ~ "Other"
    ),
    continent = factor(continent, levels = c(continent_levels,"Not found")),
    country_simple = factor(
      country_simple, 
      levels = c("United Kingdom","Europe","United States","China","Other")
    )
  )




## quick plot
MreDat_country %>%
  group_by(country_simple, year)%>%
  summarise(
    y=sum(y)
  )%>%
  ggplot()+
  geom_area(aes(x=year, y= y, fill=country_simple), position="stack")+
  theme_minimal()


```


```{r Increase efficiency - domestic freight energy efficiency}
# IEA domestic freight transport energy efficiency (gCO2/tkm)
# https://www.iea.org/data-and-statistics/data-product/energy-efficiency-indicators
eff <- readxl::read_excel(
  here::here("data/raw-data/external/IEA-shipping-energy-efficiency.xlsx"), 
  sheet = "Data") %>%
  mutate(
    y= 1/carbon_intensity_gCO2_per_tkm,
    year = as.numeric(format(as.Date(paste0(year,"-01-01")), "%Y")),
    variable_name = "Shipping carbon efficiency (tkm/gCO2)",
    component = "deployment",
    oro_type = "Efficiency"
    ) 

ggplot(eff, aes(x=year, y=y, fill=oro_type))+
  geom_col()+
  scale_fill_manual(name = "ORO type",values = typeAES$colour, breaks = typeAES$level)+
  theme_minimal()+
  theme(legend.position = "bottom")


```

**note no country level data for incr. efficiency*

For CCS download from [oil and gas climate initiative](https://www.ogci.com/ccus/co2-storage-catalogue). Choose to sum the variable of Storage Capacity: Those quantities of Total Storage Resources anticipated to be commercially accessible in the characterized geologic formation by application of development projects from a given date forward under defined conditions. Commercial Storage Resources must further satisfy four criteria: The target geologic formation must be discovered and characterized (including containment); it must be possible to inject at the required rates; the development project must be commercial; and the storage resource must remain, as of the evaluation date (i.e., not previously used for storage), on the development project(s) applied (https://www.ogci.com/ccus/co2-storage-catalogue/co2-storage-catalogue-about-the-catalogue) -- but only if associated with a project 

```{r CCS}


ccsDat <- readxl::read_excel(here::here("data/raw-data/external/CRSC_CYCLE_4_FINAL_2024_160724.xlsx")) %>%
  filter(grepl("shore|sea", area, ignore.case=T), !is.na(year_of_publication)) %>%
  filter(project_spec == "YES") %>%
  rename(year = year_of_publication) %>%
  group_by(year) %>%
  summarise(
    y = sum(sum_low, sum_mid, sum_high, na.rm=T) #sum(stored_low, stored_mid, stored_high, na.rm=T)
  ) %>%
  mutate(
    year = as.numeric(year),
    variable_name = "Storage capacity (Mt)",
    component = "deployment",
    oro_type = "CCS"
  )

ggplot(ccsDat, aes(year, y))+
  geom_col()

```


```{r CCS - country}


ccsDat_country <- readxl::read_excel(here::here("data/raw-data/external/CRSC_CYCLE_4_FINAL_2024_160724.xlsx")) %>%
  filter(grepl("shore|sea", area, ignore.case=T), !is.na(year_of_publication)) %>%
  filter(project_spec == "YES") %>%
  rename(year = year_of_publication) %>%
  group_by(year, country) %>%
  summarise(
    y = sum(sum_low, sum_mid, sum_high, na.rm=T) #sum(stored_low, stored_mid, stored_high, na.rm=T)
  ) %>%
  mutate(
    year = as.numeric(year),
    variable_name = "Storage capacity (Mt)",
    component = "deployment",
    oro_type = "CCS"
  )




# Parse country names
ccsDat_country$country <- countrycode(sourcevar = ccsDat_country$country,
                            origin = "country.name",
                            destination = "country.name")

ccsDat_country$continent <- countrycode(sourcevar = ccsDat_country$country,
                            origin = "country.name",
                            destination = "continent")


ccsDat_country$continent[is.na(ccsDat_country$continent)] <- "Not found"


ccsDat_country <- ccsDat_country %>%
  mutate(
    country_simple = case_when(
      grepl("United States", country, ignore.case=T) ~ country,
      grepl("China", country, ignore.case=T) ~ "China",
      grepl("United Kingdom", country, ignore.case=T) ~ country,
      grepl("Europe", continent, ignore.case=T) ~ continent,
      TRUE ~ "Other"
    ),
    continent = factor(continent, levels = c(continent_levels,"Not found")),
    country_simple = factor(
      country_simple, 
      levels = c("United Kingdom","Europe","United States","China","Other")
    )
  )




## quick plot
ccsDat_country %>%
  group_by(country_simple, year)%>%
  summarise(
    y=sum(y)
  )%>%
  ggplot()+
  geom_area(aes(x=year, y= y, fill=country_simple), position="stack")+
  theme_minimal()


```


Maybe can relate blue carbon to BC variable crossed with restoration

```{r CDR-BC - number of restoration projects Duarte et al 2020}
bc_restoration <- readxl::read_excel(here::here("data/raw-data/external/Duarte2020RestorationDataSet-1.xlsx"),
                                     sheet = "Data")%>%
  filter(Habitat %in% c("Mangrove","Saltmarshes","Seagrass meadows")) %>% # Omit Oyster Reefs and Coral Reefs
  # Sum the number of projects /year
  mutate(year = as.numeric(Year)) %>%
  group_by(year) %>%
  summarise(
    y = n()
  ) %>%
  ungroup() %>% 
  # # Calculate cumulative value?
  # tidyr::complete(year, fill = list(y = 0)) %>%
  # arrange(year) %>%
  # mutate(y = cumsum(y)) %>%
  mutate(
    variable_name = "BC restoration projects (N)",
    component = "deployment",
    oro_type = "CDR-BC"
  )

head(bc_restoration)

ggplot(bc_restoration, aes(x=year, y=y))+
  geom_col()
```


```{r CDR-BC - country}
bc_restoration_country <- readxl::read_excel(here::here("data/raw-data/external/Duarte2020RestorationDataSet-1.xlsx"),
                                     sheet = "Data")%>%
  filter(Habitat %in% c("Mangrove","Saltmarshes","Seagrass meadows")) %>% # Omit Oyster Reefs and Coral Reefs
  # Sum the number of projects /year
  mutate(year = as.numeric(Year)) %>%
  mutate(
    project_id = row_number()
  )



# parse country
bc_restoration_country$countries <- extract_all_affiliation(bc_restoration_country$Location, countries_ls)


# explode data frame so that its long form
bc_restoration_country <- bc_restoration_country %>%
  mutate(
    country = str_split(countries, ",\\s*"),
    country = map(country, ~ unique(str_trim(str_replace_all(.x, "^['\"]|['\"]$", ""))))
  ) %>%
  unnest(country) %>%
  filter(!is.na(country), country != "") %>%
  # add weighting 
  group_by(project_id) %>%
  mutate(
    country_weight = 1/n(),
    country_count = 1
  )

# Tabulate
bc_restoration_country <- bc_restoration_country %>%
  group_by(year, country) %>%
  summarise(
    y = sum(country_weight)
  ) %>%
  ungroup() %>% 
  mutate(
    variable_name = "BC restoration projects (N)",
    component = "deployment",
    oro_type = "CDR-BC"
  )



# Get continent
bc_restoration_country$continent <- countrycode(sourcevar = bc_restoration_country$country,
                            origin = "country.name",
                            destination = "continent")


bc_restoration_country$continent[is.na(bc_restoration_country$continent)] <- "Not found"


bc_restoration_country <- bc_restoration_country %>%
  mutate(
    country_simple = case_when(
      grepl("United States", country, ignore.case=T) ~ country,
      grepl("China", country, ignore.case=T) ~ "China",
      grepl("United Kingdom", country, ignore.case=T) ~ country,
      grepl("Europe", continent, ignore.case=T) ~ continent,
      TRUE ~ "Other"
    ),
    continent = factor(continent, levels = c(continent_levels,"Not found")),
    country_simple = factor(
      country_simple, 
      levels = c("United Kingdom","Europe","United States","China","Other")
    )
  )




## quick plot
bc_restoration_country %>%
  group_by(country_simple, year)%>%
  summarise(
    y=sum(y)
  )%>%
  ggplot()+
  geom_area(aes(x=year, y= y, fill=country_simple), position="stack")+
  theme_minimal()



```


Other CDR not available at country level. 
Make proportional to amount of available habitat area? https://essd.copernicus.org/articles/16/3433/2024/#&gid=1&pid=1 -- van Zelst et al 2022

```{r CDR-OAE CDR-BioPump}
# Ocean visions field trials 
fieldTrials <- readxl::read_excel(
  here::here("data/raw-data/external/Ocean-visions-mCDR-field-trial-database.xlsx"),
  sheet = "Data") %>%
  mutate(
    `Start of Pilot` = replace(`Start of Pilot`, `Start of Pilot` %in% c("5.12.2023","2023"), as.numeric(as.Date("2023-01-01"))),
    trialID = row_number()
  ) %>%
  mutate(year = as.Date(as.numeric(`Start of Pilot`), origin = "1899-12-30")) %>%
  mutate(year = ifelse(year == as.Date("1952-12-30"), as.Date("2023-01-01"), year)) %>%
  mutate(year = as.Date(year, origin = as.Date("1970-01-01"))) %>%
  separate_rows(`All CDR Methods`, sep=",") %>% 
  mutate(
    oro_type = case_when(
      grepl("OAE|Alkalinity|Weathering", `All CDR Methods`) ~ "CDR-OAE",
      grepl("Upwelling", `All CDR Methods`) ~ "CDR-BioPump",
      TRUE ~ "Other"
    ),
    year = as.numeric(format(year, "%Y")),
    source = "Field trial"
  ) %>%
  filter(!is.na(year) & oro_type != "Other") 
  # group_by(oro_type, year) %>%
  # summarise(
  #   y=n_distinct(trialID)
  # ) %>%
  # mutate(
  #   variable_name = "Field Trials (N)",
  #   component = "deployment"
  # )


# Founding year of mcdr startups
# Need to de-duplicate and add year  

GESAMP_companies <- readxl::read_excel(
  here::here("data/raw-data/external/GESAMP_wg41_ocean_climate_intervention_projects_31_may_2024.xlsx"), sheet="AllTables") %>%
  select(Company, Type, Website) %>%
  mutate(
    Company = trimws(gsub("[[:punct:]]", "", Company)),
    source = "GESAMP WG 41"
  )%>%
  filter(!is.na(Company))

# Load in OceanNET companies and de-deuplicate from GESAMP companies
OceanNET_companies <- read.csv(
  here::here("data/raw-data/external/OceanNETs_D18_oceanbased_CDR_companies/dataset/D1_8_database_oceanbased_companies.csv")) %>%
  mutate(
    Company = trimws(gsub("[[:punct:]]", "", Company)),
    source = "OceanNETs (2020) D1.8 database"
  )%>%
  filter(!is.na(Company), !is.na(Lat))%>%
  select(-c(contains("X"))) %>%
  filter(!(tolower(Company) %in% tolower(GESAMP_companies$Company)), Company != "Qilibrium")

## Join and write to csv file so I can look up the years on linkedIN
# # Join
# mCDRCompanies <- GESAMP_companies %>%
#   bind_rows(OceanNET_companies)
# 
# # write to file
# write.csv(mCDRCompanies, file = here::here("data/derived-data/mCDR-companies-dedup.csv"))
# 


## read in the years
mCDRCompanies <- readxl::read_excel(here::here("data/raw-data/mCDR-companies-linkedin.xlsx")) %>%
  filter(!is.na(`Company founded on (linkedin)`)) %>%
  rename(year = `Company founded on (linkedin)`) %>%
  select(Company, year) %>%
  left_join(GESAMP_companies %>% select(Company, Type), by="Company") %>%
  left_join(OceanNET_companies %>% select(Company, Type), by="Company") %>%
  mutate(
    Type = ifelse(is.na(Type.x), Type.y, Type.x)
  ) %>%
  select(Company, Type, year) %>%
  mutate(
    oro_type = case_when(
      grepl("Upwell|fertilization", Type, ignore.case=TRUE) ~ "CDR-BioPump",
      # grepl("Biomass sinking|Farming|Harvesting|Aquaculture", Type, ignore.case=TRUE) ~ "CDR-Cult",
      grepl("OAE|Alkalinity|weathering", Type) ~ "CDR-OAE",
      TRUE~"CDR-Other"
    ),
    source = "Startup",
    year = as.numeric(year)
  ) %>%
  filter(oro_type != "CDR-Other")
  
# Ocean visions community and www.cdr.fyi/leaderboards can't be scraped because they use private APIs
# Although the latter a list of suppliers can be found here without metadata
# https://www.cdr.fyi/api/search




## Combine data sources
mCDRDeploy <- fieldTrials[,c("oro_type","year","source")] %>%
  bind_rows(mCDRCompanies[,c("oro_type","year","source")]) %>%
  group_by(oro_type, year) %>%
  summarise(
    y=n()
  ) %>%
  mutate(
    variable_name = "Field trial/startup company (N)",
    component = "deployment"
  )



## make a general category
mCDR_gen <- mCDRDeploy %>%
  group_by(year, component, variable_name) %>%
  summarise(
    y = sum(y, na.rm=T)
  )%>%
  mutate(
    oro_type = "CDR-General"
  )

mCDRDeploy <- mCDRDeploy %>%
  bind_rows(mCDR_gen)

ggplot(mCDRDeploy, aes(year, y, fill=oro_type))+
  geom_col()

```

```{r CDR-OAE CDR-BioPump -- country}
library(MazamaSpatialUtils)

# Ocean visions field trials 
fieldTrials_country <- readxl::read_excel(
  here::here("data/raw-data/external/Ocean-visions-mCDR-field-trial-database.xlsx"),
  sheet = "Data") %>%
  mutate(
    `Start of Pilot` = replace(`Start of Pilot`, `Start of Pilot` %in% c("5.12.2023","2023"), as.numeric(as.Date("2023-01-01"))),
    trialID = row_number()
  ) %>%
  mutate(year = as.Date(as.numeric(`Start of Pilot`), origin = "1899-12-30")) %>%
  mutate(year = ifelse(year == as.Date("1952-12-30"), as.Date("2023-01-01"), year)) %>%
  mutate(year = as.Date(year, origin = as.Date("1970-01-01"))) %>%
  separate_rows(`All CDR Methods`, sep=",") %>% 
  mutate(
    oro_type = case_when(
      grepl("OAE|Alkalinity|Weathering", `All CDR Methods`) ~ "CDR-OAE",
      grepl("Upwelling", `All CDR Methods`) ~ "CDR-BioPump",
      TRUE ~ "Other"
    ),
    year = as.numeric(format(year, "%Y")),
    source = "Field trial"
  ) %>%
  filter(!is.na(year) & oro_type != "Other") 

# Get country and continent from coordinates
fieldTrials_country$country <- getCountry(fieldTrials_country$longitude,fieldTrials_country$Latitude)
fieldTrials_country$continent <- countrycode(
  fieldTrials_country$country, origin = "country.name", destination = "continent"
)


# Founding year of mcdr startups
# Need to de-duplicate and add year  

GESAMP_companies_country <- readxl::read_excel(
  here::here("data/raw-data/external/GESAMP_wg41_ocean_climate_intervention_projects_31_may_2024.xlsx"), sheet="AllTables") %>%
  select(Company, Type, Website) %>%
  mutate(
    Company = trimws(gsub("[[:punct:]]", "", Company)),
    source = "GESAMP WG 41"
  )%>%
  filter(!is.na(Company))%>%
  mutate(
    country = NA,
    continent = NA
  )

# Load in OceanNET companies and de-deuplicate from GESAMP companies
OceanNET_companies_country <- read.csv(
  here::here("data/raw-data/external/OceanNETs_D18_oceanbased_CDR_companies/dataset/D1_8_database_oceanbased_companies.csv")) %>%
  mutate(
    Company = trimws(gsub("[[:punct:]]", "", Company)),
    source = "OceanNETs (2020) D1.8 database"
  )%>%
  filter(!is.na(Company), !is.na(Lat))%>%
  select(-c(contains("X"))) %>%
  filter(!(tolower(Company) %in% tolower(GESAMP_companies_country$Company)), Company != "Qilibrium")

# Get country and continent from coordinates
OceanNET_companies_country$country <- getCountry(OceanNET_companies_country$Long,OceanNET_companies_country$Lat)
OceanNET_companies_country$continent <- countrycode(
  OceanNET_companies_country$country, origin = "country.name", destination = "continent"
) 



## read in the years
mCDRCompanies_country <- readxl::read_excel(here::here("data/raw-data/mCDR-companies-linkedin.xlsx")) %>%
  filter(!is.na(`Company founded on (linkedin)`)) %>%
  rename(year = `Company founded on (linkedin)`, company_country = `Company country (linkedin)`) %>%
  mutate(
    company_country = countrycode(company_country, origin="iso2c",
                                  destination = "country.name")
  )%>%
  select(Company, year, company_country) %>%
  left_join(GESAMP_companies_country %>% select(Company, Type, country), by="Company") %>%
  left_join(OceanNET_companies_country %>% select(Company, Type, country), by="Company") %>%
  mutate(
    country = case_when(
      !is.na(country.x)~country.x,
      !is.na(country.y)~country.y,
      TRUE~company_country
    ),
    country = countrycode(country, origin = "country.name",destination = "country.name"),
    continent = countrycode(country, origin = "country.name",destination = "continent"),
    Type = ifelse(is.na(Type.x), Type.y, Type.x)
  ) %>%
  select(Company, Type, year, country, continent) %>%
  mutate(
    oro_type = case_when(
      grepl("Upwell|fertilization", Type, ignore.case=TRUE) ~ "CDR-BioPump",
      # grepl("Biomass sinking|Farming|Harvesting|Aquaculture", Type, ignore.case=TRUE) ~ "CDR-Cult",
      grepl("OAE|Alkalinity|weathering", Type) ~ "CDR-OAE",
      TRUE~"CDR-Other"
    ),
    source = "Startup",
    year = as.numeric(year)
  ) %>%
  filter(oro_type != "CDR-Other")
  
# Ocean visions community and www.cdr.fyi/leaderboards can't be scraped because they use private APIs
# Although the latter a list of suppliers can be found here without metadata
# https://www.cdr.fyi/api/search




## Combine data sources
mCDRDeploy_country <- fieldTrials_country[,c("oro_type","year","source","country","continent")] %>%
  bind_rows(mCDRCompanies_country[,c("oro_type","year","source","country","continent")]) %>%
  group_by(oro_type, year, country, continent) %>%
  summarise(
    y=n()
  ) %>%
  mutate(
    variable_name = "Field trial/startup company (N)",
    component = "deployment",
    continent = ifelse(is.na(continent), "Not found", continent),
    country_simple = case_when(
      grepl("United States", country, ignore.case=T) ~ country,
      grepl("China", country, ignore.case=T) ~ "China",
      grepl("United Kingdom", country, ignore.case=T) ~ country,
      grepl("Europe", continent, ignore.case=T) ~ continent,
      TRUE ~ "Other"
    ),
    # factor
    continent = factor(continent, levels = c(continent_levels,"Not found")),
    country_simple = factor(
      country_simple, 
      levels = c("United Kingdom","Europe","United States","China","Other")
    )
  )




## quick plot
mCDRDeploy_country %>%
  group_by(country_simple, year, oro_type)%>%
  summarise(
    y=sum(y)
  )%>%
  ggplot()+
  geom_area(aes(x=year, y= y, fill=country_simple), position="stack")+
  facet_wrap(vars(oro_type))+
  theme_minimal()


```





```{r save all deployment data, eval=FALSE}
cols <- c("component","oro_type","year","variable_name","y")

allDeployDat <- rbind(
  MreDat[,cols],
  eff[,cols],
  ccsDat[,cols],
  bc_restoration[,cols],
  mCDRDeploy[,cols]
)

ggplot(allDeployDat, aes(year, y, col=oro_type))+
  geom_line()+
  facet_wrap(vars(oro_type), scales="free")+
  geom_text(aes(label = str_wrap(variable_name, 25)), x=-Inf, y=Inf, check_overlap = T, guide=FALSE, vjust=1, hjust=0, size=3)+
  labs(x = "Year", y="Action metric")+
  theme_bw()+
  theme(
    legend.position = "none"
  )



save(allDeployDat, file=here::here("data/derived-data/mitigationDeployDat.RData"))

```


```{r save all deployment data by country, eval=FALSE}
cols <- c("component","oro_type","year", 
          "country", "country_simple","continent",
          "variable_name","y")

eff_country <- eff %>%
  mutate(
    country = NA,
    country_simple = NA,
    continent = NA
  )

allDeployDat_country <- rbind(
  MreDat_country[,cols],
  eff_country[,cols],
  ccsDat_country[,cols],
  bc_restoration_country[,cols],
  mCDRDeploy_country[,cols]
)




allDeployDat_country %>%
  group_by(variable_name, oro_type, country_simple, year)%>%
  summarise(
    y=sum(y)
  )%>%
  ggplot(aes(year, y, fill=country_simple, col=country_simple))+
  geom_area(position = "stack")+
  geom_line()+
  facet_wrap(vars(oro_type), scales="free")+
  geom_text(aes(label = str_wrap(variable_name, 25)), x=-Inf, y=Inf, check_overlap = T, guide=FALSE, vjust=1, hjust=0, size=3)+
  labs(x = "Year", y="Action metric")+
  theme_bw()
  # theme(
  #   legend.position = "none"
  # )



save(allDeployDat_country, file=here::here("data/derived-data/mitigationDeployDat_country.RData"))

```


# N Posts

```{r summarise posts and save}
# before i was using n_posts_like_weighted instead of n_posts_weighted (qry + like)

postsDat <- read.csv(here::here("outputs/sentiment_predictions/sentiment_summary_oro_year.csv"))%>%
  mutate(
    weighted_ratio_pn = weighted_positive/weighted_negative,
    prop_positive = weighted_positive/n_posts_weighted
  ) %>%
  select(oro_type, year, n_posts_qry_weighted, n_posts_weighted, 
         weighted_positive, weighted_negative, weighted_ratio_pn, prop_positive)

postsDat <- reshape2::melt(postsDat,
                           id.vars = c("oro_type","year"),
                           variable.name = "variable_name",
                           value.name = "y")
postsDat <- postsDat %>%
  mutate(
    variable_name = case_when(
      variable_name=="n_posts_qry_weighted" ~ "N posts (query-weighted)",
      variable_name=="n_posts_weighted" ~ "N posts (query and like-weighted)",
      variable_name=="weighted_positive"~"N positive posts (query and like-weighted)",
      variable_name=="weighted_negative"~"N negative posts (query and like-weighted)",
      variable_name=="prop_positive"~"Proportion positive posts",
      variable_name=="weighted_ratio_pn"~"Ratio positive:negative posts",
      TRUE ~ "other"
      )) %>%
  mutate(
    component = case_when(
      variable_name=="N posts (query-weighted)" ~ "public interest",
      variable_name=="N posts (query and like-weighted)" ~ "public interest weighted",
      variable_name=="N positive posts (query and like-weighted)"~"public support",
      variable_name=="N negative posts (query and like-weighted)"~"public opposition",
      variable_name=="Proportion positive posts"~"public support (relative)",
      variable_name=="Ratio positive:negative posts"~"public support agreement",
      TRUE ~ "other"
      ),
    oro_type = gsub("_","-", toupper(oro_type))
  ) %>%
  mutate(
    oro_type = case_when(
      oro_type == "CDR-BIOPUMP" ~ "CDR-BioPump",
      oro_type == "CDR-GENERAL" ~ "CDR-General",
      oro_type == "INCR-EFF" ~ "Efficiency",
      oro_type == "MRE-GENERAL" ~ "MRE-General",
      oro_type == "MRE-BIO" ~ "MRE-Bio",
      oro_type == "MRE-LOCATED" ~ "MRE-Located",
      oro_type == "MRE-OCEAN" ~ "MRE-Ocean",
      TRUE ~ oro_type
    )
  ) %>%
  filter(variable_name != "other", oro_type != "CDR-CULT") 


unique(postsDat$component)
unique(postsDat$oro_type)

# Add MRE_general to both MREs, and CDR_general to all CDRs?
ggplot(postsDat, aes(year, y, col=oro_type))+
  facet_wrap(vars(component), scales = "free_y")+
  geom_line()

save(postsDat, file=here::here("data/derived-data/mitigationPostsDat.RData"))
```


```{r calculate summary for all mitigation OROs}

all_postsDat <- read.csv(here::here("outputs/sentiment_predictions/sentiment_allCompiled_quantCols.csv")) %>%
  distinct(post_id, .keep_all = TRUE) %>%
  mutate(
    post_date = as.Date(post_date),
    year = as.numeric(format(post_date, "%Y"))
  ) %>%
  group_by(year) %>%
  summarise(
    `N posts (query-weighted)` = sum(qry_weight),
    `N positive posts (query and like-weighted)` = sum(weight[post_sentiment == "positive"]),
    `N negative posts (query and like-weighted)` = sum(weight[post_sentiment == "negative"])
  )%>%
  pivot_longer(
    cols = c(`N posts (query-weighted)`,
             `N positive posts (query and like-weighted)`,
             `N negative posts (query and like-weighted)`
             ),
    names_to = "variable_name",
    values_to = "y"
  ) %>%
  mutate(
    component = case_when(
      variable_name=="N posts (query-weighted)" ~ "public interest",
      variable_name=="N positive posts (query and like-weighted)"~"public support",
      variable_name=="N negative posts (query and like-weighted)"~"public opposition",
      TRUE ~ "other"
      ),
    oro_type = "All mitigation OROs"
  ) %>%
  filter(component != "other", !is.na(year)) 
    
summary(all_postsDat)

save(all_postsDat, 
     file=here::here("data/derived-data/mitigationPostsDat_allMitigationOROs.RData"))
```
